Move networking and connection manager into network module
